+++ src/tol_colors/changelog.md
@@ -5,6 +5,7 @@
 - Change access: use dictionaries and make colorsets and colormaps accessible as
   module attributes. Discrete rainbow has its own dedicated function.
 - Register colormaps in matplotlib with "tol." prefix
+- Add function to automatically change Matplotlib defaults colors.
 - Fix WhOrBr not being available as discrete
 
 ## v1.4